requirements.txt
python-whois
openpyxl
